Remove update_history from automate_history.py and rename to parse_pr.py
