+++ .github/workflows/check-pr-body.yml
@@ -11,4 +11,4 @@ jobs:
       - name: Check out the repository
         uses: actions/checkout@v2
       - name: Run PR body checker
-        run: python ./scripts/automate_history.py --check-only ${{ github.event.pull_request.number }} "${{ github.event.pull_request.body }}"
+        run: python ./scripts/parse_pr.py "${{ github.event.pull_request.body }}"
